Changes index init vector
